OVN metadata agent additional_chassis detection
The patch changes how additional_chassis column support is handled in
events. We cannot call to IDL from the match methods because the post
fork event might not be set yet right after IDL was instantiated. If
between the IDL instantiation and the post fork event set an event
calling to IDL is processed, the match event method will wait
indefinitely. This patch removes the call to IDL in the match method.

Closes-Bug: #2086740
Change-Id: Ibc7d9b4dd196bed65cff73b79d78122f70aac1a7
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>