[Fullstack] Use only one keepalived-state-change thread in L3 agent
By default number of threads used by the neutron-keepalived-state-change
service is set to "1 + <num_of_cpus> / 2" which in CI results with "4".
This is definitely not needed for the fullstack tests where L3 agent is
spawned for the single test and don't need to handle more than one
router ever.
To safe some CPU resources this patch sets this config option to '1' in
fullstack tests.

Related-bug: #2083609
Change-Id: I18cfb18abe481f47db870f210188e1a570844077