[stable-only] Change the agent AZ query by agent type
In PostgreSQL, it is needed to add the field used in the query
field ("select") clause in the "group_by" one. This patch changes
the scope of the query to only select the filtered fields (availability
zones and agent type).

Because the PostgreSQL support has been dropped in master branch
(Epoxy, 2025.1), this patch is only for stable branches.

Closes-Bug: #2086787
Change-Id: Ifb5ab94ca68a9ab84407b54ac632164860b7a3a8