Catch when the process does not exist when killing it
In ``ProcessManager.disable``, it could happen that the process to
be stopped is no longer present in the system. In that case, catch
this exception and dismiss it. If the goal of the ``disable`` method
is to stop the process, it should not fail in the case of not
being present anymore.

Closes-Bug: #2088154
Change-Id: I5c6f7648d69e3a939445273f8d94241818538fc9