Catch when the process does not exist when killing it
In ``ProcessManager.disable``, it could happen that the process to
be stopped is no longer present in the system. In that case, catch
this exception and dismiss it. If the goal of the ``disable`` method
is to stop the process, it should not fail in the case of not
being present anymore.

[Stable Only] Also backport [1] to make pep8 happy for py38.
[1] https://review.opendev.org/941154
Closes-Bug: #2088154
Change-Id: I5c6f7648d69e3a939445273f8d94241818538fc9
(cherry picked from commit 0c29e730db2629c084de0c114a0d1e8e6939ac25)