Update Nova aggregates on changed host mappings
When creating a subnet on a segment, Nova aggregates are updated
with the host information. But when adding a compute node to an
existing segment or modifying what segments a node is attached to,
Nova was not updated with these changes for ML2/OVN.

ML2/OVS has agent code which via report_state() will call
create_or_update_agent() which causes the aggregates to eventually
get updated via AGENT_AFTER_CREATE events, etc.

ML2/OVN does not have "real" agents. It monkeypatches some agent
methods to respond to the API requests itself--but it does not use
the agents db--which is what create_or_update_agent() modifies.

But it shouldn't be necessary to rely on updates from the agent in
our case, as the segments code can see segment host mappings being updated and just directly handle notifying nova when those change.

Closes-Bug: #2096941
Change-Id: I8112076f8acb821752941396e7aa39ecb1352ca3