[stable-only] Remove grenade and skip-level jobs in the CI
These jobs are currently broken due to a partial migration to un/2023.1.
Because the older branch is now in unmantained mode, these jobs are
not longer executed in this stable branch.

NOTE: devstack will need [1] once this patch is merged. Because we need
to unblock our CI, this patch doesn't depend on the devstack one.

[1]https://review.opendev.org/c/openstack/devstack/+/936126

Related-Bug: #2089522
Change-Id: I0f75dbc22f2d862ed2c81b8803d8a2b4dc268ab4