Remove port resource change after unplug
Ports are removed (unplug) from the br-int, but
the ovs-agent resource_cache still have them.
If during this removal period, the port attributes
changed/updated. Even worse if the update events
were not received by the agent, as well as the
port plug to the br-int again. The ovs-agent will
use the obsolete resource informaction of the port.
Something may not be processed well then.

So, resource cache should be delete after port removed
from br-int. This patch addresses this issue.

Closes-Bug: #2089661

Change-Id: Ia628ee196382b0672ce3424d7cb86f29a15a6d4d