Drop zuul_work_dir var from neutron-functional job
[1] moved the var to neutron-functional and as that job
is used as parent by stadium projects functional job, it
resulted into those jobs to be broken.
This patch just drop this var from neutron-functional job
and if any project need to run this job directly(i.e run neutron tests),
can add the var in it's project config. Currently we running it in
neutron-lib, so that can be adjusted.

[1] https://review.opendev.org/c/openstack/neutron/+/935626

Closes-Bug: #2089673
Change-Id: Ie5b3760b716168d720bcd922ab32af11c5efbdbc