Fast exit when initially creating tunnel allocations
If the database tunnel allocations match the current configuration, no
other operation is done. During the initiallization, if several workers
try to execute this method, only the first one will update the
allocations. The next workers will check that the database is correctly
updated and will exit this method.

Closes-Bug: #2089940
Related-Bug: #2083570
Change-Id: I208ba38bff9191cabcc1325fec516d0b0179c97c