[Fullstack] Don't assert no connectivity after host is stopped
In the fullstack L3 HA tests, after fake host with router was killed
or disconnected (so wasn't stopped gracefully) there was assertion that
connectivity using Floating IP is broken for short period of time.
This could lead to the race between test and keepalived which is doing
failover. If keepalived was fast enough, test could fail at this
assertion as failover would already happen and Floating IP would be
working fine again.

This patch removes that assertion of no connectity as this can't be
really controlled by the test and may lead to random failures. It's also
not the most important thing in those tests - we should make sure that
connectivity is working fine after host is crashed thanks to the
failover which should happen and that is tested in the same tests.

Closes-bug: #2091021
Change-Id: Ib00c9823e3600bb2c234cbc90cac81723b4eec11