[OVN] Reduce the OVN hash ring touch interval
In some loaded environments, the maintenance thread in charge of
refreshing the OVN hash ring node delays the execution of the method.
While it is not possible to ensure this periodicity in loaded
environments, this patch is reducing the interval in order to trigger
the method more often.

NOTE: this change is temporary until the eventlet deprecation is
finished. With kernel threads (preemptive), the thread in charge of
refreshing the hash ring register in the database should be executed
on time. The related bug will remain open until this patch can be
reverted.

Related-Bug: #2091284
Change-Id: I5dace9cd4f1c190b647f2dc8184c466248adedd4