QinQ API extension implementation in the Neutron server
This patch implements 'qinq' API extension in the Neutron server. It
means that this new attribute is now added for the "vlan" networks to
the network dict and returned through the API.
This patch also adds validation that both "vlan_transparent" and "qinq"
aren't enabled for the same network at the same time as this is not
supported.
This patch adds all the necessary bits to store value of the new
attribute in the Neutron DB and to add support for it to the Network
OVO.
Finally it also adds check_vlan_qinq() method to all mechanism drivers
which are in-tree. For now all of them declare that QinQ vlans are not
supported.

Related-Bug: #1915151
Change-Id: I427edfd580eb06aa4f6904f90ff28cf8b5267397