QinQ implementation for the ML2/OVN backend
This patch implements support for the 'vlan_qinq' network parameter in
the ML2/OVN backend. It is done in almost similar way to the
'vlan_transparent' parameter, the difference is in the 'ethtype' set for
the provnet port for the network. For QinQ it is set to '802.1ad'.

It also adds functional tests for the 'vlan_transparent' setting for the
OVN mechanism driver. The reason why those 2 are tested together is that
both are using the same options on the OVN side and are mutually
exclusive so we have to make sure we set those options as expected in
each case.

Related-Bug: #1915151
Change-Id: I110c366a37a65d625083a7112f1adb9a3dc5e7cc