Accept both error 400 and 500 in the vlan_transparent unit tests
Error code returned from the Neutron when attempt to create network with
enabled 'vlan_transparent' has changed with [1] from 500 to 400. The
unit test which checks this has to be adjusted as well. For now, test
needs to accept both return codes as patch which changes that is in
neutron-lib and we would have "checkien and egg" issue with merging both
patches otherwise.

[1] https://review.opendev.org/c/openstack/neutron-lib/+/937605

Related-Bug: #2091613
Change-Id: Ic53b191ecfd11d71f85251a777bc0c7da4463a06