Add limit of tags for every resource
There is now limit of 50 tags which can be created for every single resource.
Limit of 50 was choosen as this should be enough for every use case.
Also e.g. Nova already have limit of 50 tags for server implemented so
with such limit Neutron will be consistent with Nova in that.

Closes-Bug: #2091410
Change-Id: I9b87d0425ac0dfdde75d2dd661f02d7124ea4f8b