Make API policies for tags to be working with resource attributes
This patch changes API policies for tags added with [1] but as "target"
to the policy.enforce() function there was only parent's id passed, not
the whole parent dictionary. Because of that policies related to tags
couldn't match on the parent's attributes, like e.g. network's "shared"
attribute.
This patch changes that so now the dict with all attributes used
potentially by the API policies is passed as target to the
policy.enforce()

Additionally this patch changes names of the actions related to the tags
in the API policy rules. Patch [1] introduced names like
"<action>_<resource_plural_name>_tags", for example
"update_networks_tags". This patch changes that to the pattern
"<action>_<resource_singular>:tags", for example: "update_network:tags"
as this is now consistent with all other actions and attributes in the
API policies in Neutron APIs.

Finally it also renames "parent" to the "obj" in the tagging extension
to not treat resources like e.g. network or port, etc. as parent of the
tag. Tag is more like attribute of that resource, not the child resource
of it.

[1] https://review.opendev.org/c/openstack/neutron/+/935883

Closes-bug: #2091493
Change-Id: I665ed178e4a2e01d7f94cac6b9d3b482c3ed17a8