[eventlet-removal] Use non-eventlet metadata proxy in OVN metadata agent
The new metadata proxy server implementation, used in the OVN agent, is
now used in the OVN metadata agent.

The same limitation applies: it is only possible to create a server inside
a thread, running in the same process (same as with metadata_workers=0).
The configuration variable metadata_workers no longer applies.

Closes-Bug: #2087941
Change-Id: If5d74d6e21a22e4f4dbed27ed9be828adad857e8