Make sure that policy enforcer is initialized before use
In some cases policy.enforce() or policy.check() functions were called
without calling policy.init() before and that could lead to errors like
mentioned in the related bug.
To avoid that and to make it less error prone in future, this patch
moves call of the policy.init() method directly to the enforce() and
check() methods and removes it from all other places. Other modules
should not need to carry about policy initialization at all.

Closes-Bug: #2092659
Change-Id: Ic11992ba3ed91980189efbacdc2a54fba64fcf7c