Allow plugins to add periodics to maintenance worker
Some plugins may want to run additional periodic tasks in the
maintenance worker. Extend the initialization of the maintenance thread
to also add periodics objects found from plugins.
A plugin that wants to add its own periodics will need to implement
a new function `ovn_maintenance_periodics` that is expected to return
a list of periodics objects to be added.
Additionally the OVN SB IDL (BaseOvnSbIdl) used by the
MaintenanceWorker is extended to allow adding watches.

The intended use case is the ovn-vpnaas plugin in neutron-vpnaas.
It needs to periodically check the VPN agent list to detect dead
agents and potentially reschedule VPN services to alive agents.
The notify_handler added to BaseOvnSbIdl is needed there to
update the AgentCache, at least for the VPN agents.

Related-Bug: #2089250
Change-Id: Ifd9a691926a9ae7953935bbe0b65da9e8fe8c02b
Signed-off-by: Bodo Petermann <b.petermann@syseleven.de>