[eventlet-removal][OVN] Require wsgi start-time in the config
Currently there is one single method to run the Neutron API that is
using uWSGI. This method requires a specific hash ring manager
initialization, different to the eventlet server based.

Because the second has been removed and is no longer available, as
long as Neutron continues the effors of removing eventlet from the
source code, the second way to initialize the hash ring manager
is removed.

Related-Bug: #2083570
Change-Id: I59b25093df9cf5aa77767492a5b9008bfa11cc07