[Grenade] Explicitly enable correct neutron services in devstack config
It seems that when neutron is enabed in the Devstack's services as
"enable_service q-svc" it may be started properly by Devstack but when
it comes to stopping it, it tries to stop services expected to be
enabled when it is deployed under mod_wsgi and that fails.

So this patch enables correct services in the devstack's configuration
so that those services can be started properly.

Depends-On: https://review.opendev.org/c/openstack/devstack/+/940438
Depends-On: https://review.opendev.org/c/openstack/devstack/+/940447
Depends-On: https://review.opendev.org/c/openstack/grenade/+/940579

Closes-bug: #2096912
Change-Id: Ia077097c11a1ccc9e03d5e4ceed068790c5e8758