OVS: Set log level to INFO for unconfigured ofport cases
Log an ERROR if the VIF port has no ofport, which indicates
that the port might not be able to transmit traffic. This can
happen if the ovs-vswitchd service is down, but the ovsdb-server
service is still running. In such cases, the port record might
be created in the database, but it won't be processed by
ovs-vswitchd, resulting in an empty list ofport value.

I was unable to reproduce the ofport=[] scenario using the
OpenStack CLI. However, I managed to achieve this by using the
«ovs-vsctl add-port» command while the ovs-vswitchd service was
stopped. Reusing this port in Neutron is impossible. I believe
this situation is highly unlikely to occur in practice, but I
decided to leave the check as it is, just in case.

When the ofport is set to INVALID_OFPORT, it indicates that
the port is in a transitional state and has not yet been fully
configured. This is not an error condition but rather a normal
part of the port's lifecycle, as it may take some time for the
port to be initialized and assigned a valid ofport. Operators
should be aware of this state for monitoring purposes, but no
immediate action is required.

Closes-Bug: #2095576
Change-Id: I0b7e7e8c506a9632aac911c1b4a0d19653dd5239
(cherry picked from commit a81aa17471a44545f71a044fb0008dc576be2f03)