[OVN] Change ``PeriodicWorker`` executor to ``ThreadPoolExecutor``
The default executor factory for the ``periodics.PeriodicWorker``
instance is ``futurist.SynchronousExecutor``. That creates a caller that
is executed synchronously with the current thread. The
``futurist.ThreadPoolExecutor`` instead creates a new thread pool to
execute the calls asynchronously. Now eventlet is removed from the
Neutron API and kernel threads are using, this executor will be more
accurate when calling the periodic methods.

Closes-Bug: #2097257
Change-Id: Ie01fd776856463593e016a0a22a29cb5666ea896