[OVN] Create a deterministic hash ring node UUID generator
The hash ring node UUID is now created using a deterministic generator.
This method requires the group name, the host name and the worker ID.
These three parameters are always the same, during the same execution,
for a uWSGI worker. If the worker is restarted by uWSGI, the UUID
created will be the same.

Change-Id: I8007647c2db9fa3d75aba16aa293d0d2c7e5d4c5