ml2/sriov: bind port on segment that subnet belongs
Follow up to change I56b22820d29b2d57faf28a2f9b685ab0b2c924b4

Currently, if an SRIOV agent has access to multiple segments
within a single routed provider network, the driver will always
attach to the first segment it has access to. If a port has been
explicitly associated with a subnet and hence a segment, this
behaviour may not be correct.

This patch makes use of the same feature now available in the
base driver class and ensures that where a port is already
associated with a subnet, the SRIOV driver attaches to the
correct segment.

Closes-Bug: #2098085
Change-Id: I40a0de80ebf4ba2b32f0105a7b2d533ba13bc692