Revert "Track all interfaces in Keepalived"
This reverts commit bee07defacf61ccef19adbe54677b0a0a00ae65b.

Reason for revert:
The only interfaces tracked by keepalived are the HA interfaces. The
fixed IP / floating IPs / routes are linked to the internal router interfaces or the gateway interface, that are interfaces not tracked
by keepalived. These IP addresses / routes should have the suffix
"no_track" in the configuration entry.

This is commented in [1], when the keepalived VIP HA configuration
was fixed, excluding any IP address other than the VIP of the HA interface, that are placed in the "virtual_ipaddress_excluded" and
belong to no tracked interfaces.
    
[1]https://github.com/openstack/neutron/blob/ad628353da68dcbf9e6ff25c47fd6753eee22683/neutron/agent/linux/keepalived.py#L271-L281

Change-Id: I4dfd89606042ba545559eb03d47fceee3b0895fc
Closes-Bug: #2097770