Policy validator "type:list_of_dict*" also contains sub attributes
The policy engine retrieves the API definition validators to build the
resource rule match string. If the validator starts with "type:dict",
it will contain a set of sub attributes that should be parsed too from
the API call.

This patch introduces a new set of validators, starting with
"type:list_of_dict". In this case, the main attribute will be a list
of dictionaries, instead of a single dictionary. When building the
match string, the engine will loop over the list of dictionary keys
and add them to the attributes to be present in the match string.
It could be possible that not all dictionaries have the same keys.

The sibling change needed in ``neutron-lib`` is [1]. This change is
adding this validator for ``router.external-gateways`` field, that
is a list of dictionaries.

[1]https://review.opendev.org/c/openstack/neutron-lib/+/942064

Related-Bug: #2098109
Change-Id: Ie7e6ff26a25bd3ea2b1575924162082cdbc492bf