[OVN][QoS] Don't change the input parameter in ``_add_port_qos_rules``
The method ``OVNClientQosExtension.update_policy`` iterates over the
ports in a network and updates the new QoS policy rules on them. This is
done calling the method ``OVNClientQosExtension._add_port_qos_rules``
that retrieves the QoS rules in the parameter ``qos_rules``, that is
a dictionary. This value is used in the loop that updates all the
Logical_Switch_Ports. If this dictionary is modified, the next calls
will contain wrong information.

Closes-Bug: #2099706
Change-Id: Iaa94310d008be84e3a71441b24afa8a8df9d87eb