[FT] Refactor the ``TestNeutronServer._check_active`` method.
The method ``TestNeutronServer._check_active`` now actively checks
if the workers are ready after spawning them. The method ``_fake_start``
writes a dummy value in a specific queue that is read by the mentioned
method. If the number of values equals the number of workers, then
is assumed that all workers are ready.

Closes-Bug: #2100001
Change-Id: I4f700db665784144d0e1ded1912a03ea49498ae4