[OVN] Make the Logical_Switch_Port UP event more robust
In ML2/OVN it is possible that the OVN databases are configured in a
OVSDB RAFT cluster. If by any circumstance the RAFT leader changes, the
active IDL connections to OVN NB and SB databases will temporarily
disconnect and reconnect again. During this time gap, if the
``LogicalSwitchPortUpdateUpEvent`` arrives, it won't be processed.

This patch keeps the ``LogicalSwitchPortCreateEvent`` in the watch list.
Now if the IDL reconnects, all Logical_Switch_Port registers will be
received as CREATE events. If a ``LogicalSwitchPortUpdateUpEvent`` was
missed, now it will be processed.

Closes-Bug: #2101140
Change-Id: Ie303ec230c1b0510d41289ad17ca3d3a48d7b144
(cherry picked from commit 1c953a0bcc0ddce0580055b89e006589bc92ff4a)