[OVN] Do not call ``QoSAddCommand`` without max-bw/DSCP rule
The ML2/OVN QoS extension formats the Neutron QoS rules into the
corresponding QoS registers and LSP.options values. The OVN QoS
registers allow to define filters for a specific traffic, matching
the direction and the source port. This registers can have
a "rate" and a "dscp" values defined. It is mandatory to at least
define one of these parameters. If, for example, the Neutron rules
only define a min-bw rule, none of these parameters will be set in
the dictionary with the OVN QoS rules. In that case, ``qos_add`` is
not called.

Closes-Bug: #2101063
Change-Id: Ia4f1d61753594c54f55bd146ad6dc06195a01485