[OVN] Retrieve the HA_Chassis_Group in ``_sync_ha_chassis_group``
Instead of calling the method ``ha_chassis_group_add`` that is not
returning the existing HA_Chassis_Group register if present, this patch
is first retrieving it from the database. If the register is not
present, then a creation call is issued.

Conflicts:
    neutron/tests/unit/services/ovn_l3/test_plugin.py

Closes-Bug: #2100505
Change-Id: I8b190f41e8a83e47ea64b1b2962fbca49fbc5f63
(cherry picked from commit 978d0b7bfbe156098fc6ed017a6fb0e02483a268)