[S-RBAC] Fix policies for the SG rules API
This patch fixes default policies for the Security Group Rules API so
that user of the project who isn't owner of the SG but only sees it
as shared one, can't now create or delete rules in such SG.

Additionally this patch lowers numer of retries when parent object's id
is looked up in the DB by the OwnerCheck policy rule to just one. If it
will fail twice with NotFound exception, then there is no need to repeat
it more times.

Closes-bug: #2101150
Change-Id: I23722d0ffabce0034548a5fa919980d02bacd91a