[OVN] Delete a resource provider removed from the agent config
In ML2/OVN, The Placement information per compute node is stored in the
local OVS database, in the Open vSwitch register. Each time this
information changes, ML2/OVN receives this event and updates the OVN
agent child registers assigned to devices.

Since this patch, if a resource provider is removed, the corresponding
Placement resource provider register (and its inventory) is removed too.

NOTE: if a resource provider has allocations, the delete operation will
fail with HTTP 409 (Conflict) and the Neutron API will log this error.
The error message will be:
  Unable to delete resource provider <UUID>: Resource provider has
  allocations.

Closes-Bug: #2101998
Change-Id: I78319cc3c94bf869383ecfaf14cb276b46086b8e