[OVN] Do not delete twice the agent from the cache
When a delete agent command is issued, the API worker that attends
the API call, it issues an OVN SB ``SB_Global`` event to inform any
other worker about the deletion of the agent, thus they can delete
the agent from their local OVN agent cache.

The problem is that this event is also received by the worker that
is sending it. Because the call that is generating the event is also
deleting the agent [1], this worker tries to delete the agent twice;
the second time, it generates a KeyError exception in the local OVN
agent cache.

The code added in [1] should be deleted.

[1]https://review.opendev.org/c/openstack/neutron/+/883607/3/neutron/plugins/ml2/drivers/ovn/mech_driver/mech_driver.py#1421

Closes-Bug: #2102645
Change-Id: Ie391af5a3b226cbd33f7162a3c970d636fcdf1da