Extend port resource request only when using qos minimum rules
In the method [1], if there is no minimum bandwidth rule and
minimum packet rate rule in the port's qos policy,
port['resource_request'] is set to None in the end, which means
no furthermore resource_request extension is required in those cases.

This change checks first if there are minimum bandwith rules and
minimum packet rate rules in the policy. If there is no minimum rule,
set port_res['resource_request'] to None, and then continue the
for loop.

This change adds a test checking resource_request extension is not
done and ports' resource_request is set to None in this case.

There is change in existing test. This change fixes test which is
affected by the modified method[1]. This change adds new member
variables network_id and ports_res, and uses them instead of defining
new ones in the other tests.

[1] https://github.com/openstack/neutron/blob/master/neutron/services/qo
s/qos_plugin.py#L276

Closes-Bug: #2103412
Change-Id: I4d84cbf21be5e065a6ff57ba87352e2b602862b5