Revert "[eventlet-removal] Remove the usage of eventlet in the DHCP agent"
This reverts commit 7c32d43c4b37ec7629b1ff191e54c389c03b6f8d.

Reason for revert: the DHCP agent is still using "oslo.service". The
only backend implemented so far is "eventlet" (see c#2 in LP#2101840).
This revert is temporary until we have a "oslo.service" release
using kernel libraries.

This patch is reverted along with [1], that is proposed on top of this
patch.

[1]https://review.opendev.org/c/openstack/neutron/+/942393

Partial-Bug: #2101840
Change-Id: Ifd4f10a65935d9779ae4b5329258aef6bd5f871b