[stable-only] Fix the Neutron milestones list
The patch [1] incorrectly added ``RELEASE_2023_1`` to the
``NEUTRON_MILESTONES`` list, instead of ``ZED``

The patch [2] incorrectly added ``RELEASE_2023_2`` to the
``NEUTRON_MILESTONES`` list, instead of ``RELEASE_2023_1``

The patch [3] didn't add a new release because ``RELEASE_2023_2``
was already in the list.

The patch [4] didn't add ``RELEASE_2024_1`` to the list.

The patch [5] didn't add ``RELEASE_2024_2`` to the list.

[1]https://review.opendev.org/c/openstack/neutron/+/859111
[2]https://review.opendev.org/c/openstack/neutron/+/876051
[3]https://review.opendev.org/c/openstack/neutron/+/895155
[4]https://review.opendev.org/c/openstack/neutron/+/912761
[5]https://review.opendev.org/c/openstack/neutron/+/929351

Change-Id: Ie3b4197d677cc74371acd1c73b0ae079c6eeb057