Remove flaky batch notifier unit test
The test_queue_event_multiple_events_notify_method()
can be flaky depending on the performance of the SUT,
where sometimes the notifier is called thrice instead
of the expected twice, making in unreliable. As we are
already testing all the notification events arrived in
test_queue_event_multiple_events_callback_method(),
let's remove the other test.

Closes-bug: #2100806
Change-Id: I193b4cba5b31ff9baa90992a11db70e5b65ec569