Fix some pylint warnings in the test tree
Since we don't run pylint in the test tree, things creep
back in. Fixed a number of warnings:

    expression-not-assigned
    super-with-arguments
    useless-object-inheritance
    trailing-comma-tuple
    undefined-loop-variable
    cell-var-from-loop
    used-before-assignment
    unsubscriptable-object
    pointless-statement
    confusing-with-statement

Also took the chance to address all the following types
of warnings:

    no-value-for-parameter

These were in cases where tests were using a decorator
like @mock.patch or @mock.patch.object, which can confuse
pylint. Luckily you can add a list of signature-mutators
for it to ignore, which is done in a couple of other
repos.

While we are not able to enable pylint for the test tree
quite yet, we are down to ~180 warnings, most being
unnecessary-lambda (~100) and unnecessary-dunder-call (~30).

TrivialFix

Change-Id: I94031c639451b3b4ac00f93052d3a2cd08e08fac