Add the utility class ``ThreadPoolExecutorWithBlock``
This class inherits from ``futures.ThreadPoolExecutor`` and provides
a blocking submit method that only returns when the work queue is
empty; that means, it is only possible to add a new task only if no
other tasks are waiting. This behaviour replicates the
``eventlet.GreenPool.submit`` method and is expected to replace it.

Partial-Bug: #2087943
Change-Id: I77c258cdcf3cff6ae5fdc733269c2276737971b0