support custom network settings per domain for edns logging and DNS
This adds support for configuring custom upstream DNS servers and to
disable the edns logging for neutron networks, based on the name of
the OpenStack domain or, for development, the project id.

By changing the upstream DNS servers used by dnsmasq, local hostnames
are still resolvable, but queries for external names will be sent to
the custom set of resolvers from the configuration. This is in contrast
to the existing dns server setting for networks, which changes the
nameservers announced via DHCP.

The configuration is done in neutron server.

When a dhcp-agent first configures a network, it does an rpc
request to retrieve all necessary configuration data. On the server
side we add the custom DNS servers and a flag for the edns logging to
the returned data. The agent code then uses this data to configure the
dnsmasq instances accordingly. When that data is absent the local
configuration is used as default, like before.

To decide if the network is to be configured with a custom set of
nameservers, the neutron rpc server will query keystone for the domain
of the network and match the domain name against its configuration file.

To help development and debugging, support to directly match the id
of a project was also added.