Make sure OVN PG is present before adding a SG rule
It's possible to have a security group created on one
controller and then a security group rule created on a
different controller quickly enough that the second
controller does not yet see that security group in its
local cache of the OVN northbound database. Check if
the port group is present or not in the idl's local copy
of the database before creating the security group rule.

Closes-bug: #2090921
Change-Id: If6a60eb086b9bcdfd90e154879561ba569f45750
Signed-off-by: Brian Haley <haleyb.dev@gmail.com>