Add persist_uuid upgrade scenario test
This just adds a test to make sure that a Logical_Switch added
before using persist_uuid can still have a port successfully
added to it when using persist_uuid. As long as code looking up
the Logical_switch does the lookup by name, which AddLSwitchPort
and the rest of the neutron code do, things should work.

Change-Id: I780b4631e5826aceba15406ebbcb06ca55a9ea2b