Increase threshold in TestHAL3Agent fullstack test
The HA router failover fullstack test has a maximum threshold of ICMP
packets lost after a failover has been executed. Based on the number of
failures observed in zuul recently, this threshold seems to be set at a
value too low. In this change, an increase in this value is proposed.

Change-Id: I4555ed821865db724c4e5fc502648fe71d5febea
Partial-Bug: #2104019
Partial-Bug: #2091351