Subnet filter by "router:external" needs to be changed to "external"
The extension "subnet-external-network" added a new API field to the
subnet resource: "router:external". This name was used to maintain a
coherence with the network resource field, that has the same name.

However, the Neutron subnet OVO defines this synthetic field as
"external". This patch changes how the ``Subnet.get_objects`` method
filters by "router:external"; the field name is converted to "external".

This patch is being tested in [1].

[1]https://review.opendev.org/c/openstack/neutron-tempest-plugin/+/926508

Partial-Bug: #2076328
Related-Bug: #2051831
Related-Bug: #2082883
Change-Id: I3728aa3932931320d9ccbc24fb9f8b3533dffff7