[OVN] Only update the MTU of the router GW LRPs
When a new internal network is added to a router, the GW LRP (or LRPs)
MTU value is updated, matching the lowest MTU of all networks connected
to this router.

This patch is updating the method ``set_gateway_mtu`` to be called only
for a specific router and update only the router GW LRPs.

Closes-Bug: #2109388
Change-Id: Ic7807a440b479c0fa08c6d13813cd2a81eff9bb1